+++ src-tauri/src/game_board/mod.rs
@@ -6,7 +6,7 @@ use blank_cell::BlankCell;
 
 #[derive(Clone, Copy)]
 pub enum Cell {
-    Printed(u8),
+    Printed(usize),
     Blank(BlankCell),
 }
 
@@ -17,7 +17,7 @@ pub struct GameBoard {
 impl GameBoard {
     // delete target in a cell's candidate list
     // and mark it as user deleted if user_deleted_flag is true
-    fn delete_candidate(&mut self, x: usize, y: usize, target: u8, user_deleted_flag: bool) {
+    fn delete_candidate(&mut self, x: usize, y: usize, target: usize, user_deleted_flag: bool) {
         if let Cell::Blank(cell) = &mut self.grid[x][y] {
             cell.modify(|candidates, user_deleted| {
                 if candidates.contains(target) {
@@ -31,7 +31,7 @@ impl GameBoard {
     }
 
     // check if this cell collides with the target
-    fn check_cell_collision(&self, x: usize, y: usize, target: u8) -> Option<(usize, usize)> {
+    fn check_cell_collision(&self, x: usize, y: usize, target: usize) -> Option<(usize, usize)> {
         match &self.grid[x][y] {
             Cell::Printed(ans) if target == *ans => Some((x, y)),
             Cell::Blank(blank_cell) if blank_cell.check_collision(target) => Some((x, y)),
@@ -41,7 +41,7 @@ impl GameBoard {
 
     // take an cell at (x,y) and return the vector of coordinates that collide with the target
     // the caller should ensure that the cell is a blank cell
-    fn check_clue_collision(&self, x: usize, y: usize, target: u8) -> Vec<(usize, usize)> {
+    fn check_clue_collision(&self, x: usize, y: usize, target: usize) -> Vec<(usize, usize)> {
         Coord::seeable_cells(x, y)
             .filter_map(|(xi, yi)| self.check_cell_collision(xi, yi, target))
             .collect()
@@ -52,9 +52,10 @@ impl GameBoard {
         &self.grid[x][y]
     }
 
+
     // check if the given target shouldn't be deleted
     // the caller should ensure that the cell is a blank cell
-    pub fn check_pencil_mark_deletion_error(&self, x: usize, y: usize, target: u8) -> bool {
+    pub fn check_pencil_mark_deletion_error(&self, x: usize, y: usize, target: usize) -> bool {
         matches!(self.grid[x][y], Cell::Blank(ref cell) if target == cell.get_answer())
     }
 
@@ -64,21 +65,21 @@ impl GameBoard {
         &self,
         x: usize,
         y: usize,
-        target: u8,
+        target: usize,
     ) -> Vec<(usize, usize)> {
         self.check_clue_collision(x, y, target)
     }
 
     // erase an pencil mark in given cell by user
     // the caller should ensure that the cell is a blank cell with no pen mark and the target is in the candidate set
-    pub fn erase_pencil_mark(&mut self, x: usize, y: usize, target: u8) {
+    pub fn erase_pencil_mark(&mut self, x: usize, y: usize, target: usize) {
         self.delete_candidate(x, y, target, true);
     }
 
     // add an pencil mark in given cell by user
     // and remove the user deleted flag
     // the caller should ensure that the cell is a blank cell with no pen mark and the target is not in the candidate set
-    pub fn add_pencil_mark(&mut self, x: usize, y: usize, target: u8) {
+    pub fn add_pencil_mark(&mut self, x: usize, y: usize, target: usize) {
         if let Cell::Blank(cell) = &mut self.grid[x][y] {
             cell.modify(|candidates, user_deleted| {
                 if !candidates.contains(target) {
@@ -95,7 +96,7 @@ impl GameBoard {
         &self,
         x: usize,
         y: usize,
-        target: u8,
+        target: usize,
     ) -> Option<Vec<(usize, usize)>> {
         match &self.grid[x][y] {
             Cell::Blank(cell) if target != cell.get_answer() => {
@@ -107,7 +108,7 @@ impl GameBoard {
 
     // set a cell as pen mark by user
     // the caller should ensure that the cell is a blank cell with no pen mark
-    pub fn add_pen_mark(&mut self, x: usize, y: usize, target: u8) {
+    pub fn add_pen_mark(&mut self, x: usize, y: usize, target: usize) {
         if let Cell::Blank(cell) = &mut self.grid[x][y] {
             cell.set_pen_mark(target);
 
@@ -175,7 +176,7 @@ mod game_board_test {
             let i = index / 9;
             let j = index % 9;
             if c.is_digit(10) {
-                grid[i][j] = Cell::Printed(c.to_digit(10).unwrap() as u8 - 1);
+                grid[i][j] = Cell::Printed(c.to_digit(10).unwrap() as usize - 1);
             }
         }
         GameBoard { grid }
