+++ src-tauri/src/solvers/solver_result/confirmation.rs
@@ -5,6 +5,6 @@
 pub struct Confirmation {
     pub x: usize,
     pub y: usize,
-    pub target: u8,
+    pub target: usize,
 
 }
